fix(core/llm,core/status): move normalize_status to core/ to fix cross-package import
`core/llm/response_validation._normalise_status_field`
imported `normalize_status` from
`packages.exploitability_validation.orchestrator`.
That's a layering inversion — `core/` is the lower
layer, `packages/` builds on it. The inline import
deferred the issue at import time, but it remained a
runtime cross-package coupling that:

* Broke clean dependency analysis (can't reason about
  `core/llm` without also knowing about
  `packages.exploitability_validation`)
* Would have hit a circular import the moment another
  core/ consumer also wanted the helper (and the
  validation package imports plenty of core/ modules
  during its own load)
* Made the validation package a transitive load
  dependency of every LLM-response normalisation call —
  even from tools that don't run validation
* Couldn't be unit-tested for `core/llm` without
  installing the validation package's dependencies

Move `_STATUS_ALIASES`, `normalize_status`, and
`normalize_findings` to a new `core/status` package.
Pure data + string normalisation, no LLM-specific or
validation-specific logic. Both consumers now import
from the same canonical location.

Back-compat: `packages/exploitability_validation/orchestrator`
re-exports the three names so existing callers that
import `from ...orchestrator import normalize_status`
keep working unchanged.

`core/status` is a package directory (per
`feedback_core_no_bare_files` — every new core/ unit
is a package, not a flat .py file), with `__all__` set
to the canonical names so wildcard imports stay clean.

Tests: `core/llm/tests/test_response_validation` +
`packages/exploitability_validation/` -> 405 passed.

Closes wave1-LLM W1-LLM-StatusModule (cluster 097).
