+++ packages/exploitability_validation/orchestrator.py
@@ -55,83 +55,12 @@ class StageStatus(Enum):
     SKIPPED = "skipped"
 
 
-# Alias map: any legacy/mixed-case variant → canonical snake_case.
-# Add entries here when upstream producers use a different convention.
-_STATUS_ALIASES = {
-    # ALL_CAPS legacy (orchestrator pre-cleanup, LLM skill output, docs)
-    "EXPLOITABLE": "exploitable",
-    "CONFIRMED": "confirmed",
-    "CONFIRMED_CONSTRAINED": "confirmed_constrained",
-    "CONFIRMED_BLOCKED": "confirmed_blocked",
-    "CONFIRMED_UNVERIFIED": "confirmed_unverified",
-    "RULED_OUT": "ruled_out",
-    "NOT_EXPLOITABLE": "unlikely",
-    # Title-case legacy (old feasibility verdicts, LLM output)
-    "Exploitable": "exploitable",
-    "Confirmed": "confirmed",
-    "Ruled Out": "ruled_out",
-    "Disproven": "disproven",
-    "Not disproven": "not_disproven",
-    "Likely exploitable": "likely_exploitable",
-    "Difficult": "difficult",
-    "Unlikely": "unlikely",
-    "Unknown": "unknown",
-    "Likely": "likely_exploitable",
-    # Passthrough — already canonical
-    "exploitable": "exploitable",
-    "confirmed": "confirmed",
-    "confirmed_constrained": "confirmed_constrained",
-    "confirmed_blocked": "confirmed_blocked",
-    "confirmed_unverified": "confirmed_unverified",
-    "ruled_out": "ruled_out",
-    "likely_exploitable": "likely_exploitable",
-    "difficult": "difficult",
-    "unlikely": "unlikely",
-    "unknown": "unknown",
-}
-
-
-def normalize_status(value: Optional[str]) -> Optional[str]:
-    """Normalize any status/verdict string to canonical snake_case.
-
-    Handles ALL_CAPS, Title Case, and snake_case inputs.
-    Unknown values are lowercased with spaces replaced by underscores.
-    Non-string inputs are coerced to string first.
-    """
-    if not value:
-        return value
-    if not isinstance(value, str):
-        value = str(value)
-    value = value.strip()
-    if not value:
-        return None
-    canonical = _STATUS_ALIASES.get(value)
-    if canonical:
-        return canonical
-    # Fallback: lowercase, replace spaces/hyphens with underscores
-    return value.lower().replace(" ", "_").replace("-", "_")
-
-
-def normalize_findings(data: dict) -> None:
-    """Normalize all status/verdict fields in a findings dict in-place."""
-    for finding in data.get("findings", []):
-        if not isinstance(finding, dict):
-            continue
-        if finding.get("status"):
-            finding["status"] = normalize_status(finding["status"])
-        if finding.get("final_status"):
-            finding["final_status"] = normalize_status(finding["final_status"])
-
-        ruling = finding.get("ruling")
-        if ruling and ruling.get("status"):
-            ruling["status"] = normalize_status(ruling["status"])
-
-        feasibility = finding.get("feasibility")
-        if feasibility:
-            if feasibility.get("verdict"):
-                feasibility["verdict"] = normalize_status(feasibility["verdict"])
-            if feasibility.get("status"):
-                feasibility["status"] = normalize_status(feasibility["status"])
+# `normalize_status` / `normalize_findings` moved to `core.status`
+# so `core/llm/response_validation` can import without crossing
+# layers (core → packages was a layering inversion). Re-exported
+# here for back-compat with consumers that import from the
+# orchestrator path.
+from core.status import _STATUS_ALIASES, normalize_findings, normalize_status
 
 
 # Human-readable display for status values (per CLAUDE.md style guide)
